refactor : platform img 가져오기 리팩토링
